requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.14
django-filter==24.2
djangorestframework==3.15.2
psycopg2-binary==2.9.9
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1